isnan implementation for SIMD types (#42)
* Implement isnan for SIMD types

* Add unit test for isnanSIMD

* Use xsimd version 7.5.0 for CI

* Minor fixes

* Try to fix isnanSIMD on Linux

* More isnan fixes on GCC

+++ DSP/WDF/r_type.h
@@ -47,7 +47,7 @@ namespace WDFT
 
             int size() const noexcept { return arraySize; }
 
-            alignas (alignment) std::array<ElementType, arraySize> array;
+            alignas (alignment) std::array<ElementType, (size_t) arraySize> array;
         };
     } // namespace rtype_detail
